refactor(vtscore): move evt_mixture out of the shipped training surface (#3396)
`vtscore/training/evt_mixture.py` is the Gumbel/Normal score mixture from the
the only module in `vtscore/training/` with no production path: every sibling
(`mlp`, `svm`, `thresholds/`, `blend_schedules`, `region_similarity`,
`structural_similarity`) is reachable from `vtsearch/routes/sorting.py` or
`vtscore/detectors/`, while nothing in the app or the library ever fits a
Gumbel. Its only library consumers are the eval harness's cut families
(`cut_rules`, `fit_quality`, `arms_safe_gmm`, `voting_columns`) plus the
calibration studies under `scripts/experiments/calibration/`.

Sitting beside the `thresholds/` package made it indistinguishable from the
learned-sort toolkit `vtscore/docs/packages/training.md` invites library
consumers to use, so it moves to `vtscore/eval/` — the tier that is already, by
construction, the measurement harness — and lands next to the four modules that
use it. A module banner records why it lives there.

No compat shim at the old path, on purpose: `vtscore.training` re-exporting from
`vtscore.eval` would invert the package layering and pull the eval package's
matplotlib import into the training tier. The move is recorded as a break in
`vtscore/CHANGELOG.md`; every symbol keeps its name, so an out-of-tree importer
updates one line.
