+++ vtscore/CHANGELOG.md
@@ -77,6 +77,25 @@ instead, since every commit on `dev` is effectively a new app release.)
 
 ### Changed
 
+- **`vtscore.training.evt_mixture` moved to `vtscore.eval.evt_mixture`** (issue
+  #3396). The Gumbel/Normal score mixture is a research arm from the #2836 /
+  #2846 cut studies, not a shipped threshold rule: nothing in the app or the
+  library's production path fits a Gumbel, and its only consumers are the eval
+  harness's cut families and the calibration studies under
+  `scripts/experiments/calibration/`. Sitting in `vtscore/training/` put it
+  beside the `thresholds/` package that *is* offered to library consumers, with
+  nothing to tell the two apart.
+
+  **For library consumers:** this is a breaking move of a module path, taken
+  deliberately rather than shimmed. `GumbelNormalFit1D`,
+  `fit_gumbel_normal_mixture_state`, `gaussian_mixture_mean_loglik`,
+  `CROSSING_REASONS` and the rest are unchanged and keep their names; only the
+  module they live in changed, so an importer updates one line. There is no
+  re-export at the old path on purpose: `vtscore.training` importing from
+  `vtscore.eval` would invert the package layering and drag the eval package's
+  matplotlib import into the training tier. The Gaussian threshold rules that
+  production actually ships (`vtscore.training.thresholds`) are untouched.
+
 - **A broken `vtscore.datasets` install now fails loudly instead of silently
   disabling origin resolution** (issue #3397).
   `vtscore.detectors.resolver` used to defer its imports of
